LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/moving-between-documents/ordering/in-order-expected.txt

CONSOLE MESSAGE: Not executing script because it moved between documents during fetching


PASS Script elements (in-order) still block subsequent in-order scripts in the original Document after moved to another Document
